fix: embed footer text as a constant variable, added constants.py
